Avoid using InstallValue on non-plain objects
The pair DeclareGlobalVariable / InstallValue in GAP has unfixable
design issues and arguably should never have been added in the
first place, but that ship has sailed.

The main issue with these functions is with non-plain objects
being used as value. In practice that means using values which are
plain lists, plain records or string objects is acceptable (though
still undesirable). But using a component or positional objects
requires a change of the internal representation of the
placeholder that e.g. prevents safe use in a multi threaded GAP.
Future GAP versions will therefore issues warnings when code
uses InstallValue to do that.

Luckily in many cases it suffices to use BindGlobal. In some cases
one needs to make the name of the variable known to the system
before declaring it. For that, one can use DeclareGlobalName
(introduced in GAP 4.12). This only leaves code that really tris
to access the object before an actual value has been set. Most
such uses are dangerous. In the case of this package, it only
affected a DeclareSynonym call, where it is safe in plain GAP but
not in multi-threaded GAP.